pyproject.toml
[tool.poetry]
name = "web3devlearn"
version = "0.1.0"
description = ""
authors = ["Alexander Semenov <alexander.semenov@tum.de >"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
black = "^22.12.0"
py-solc-x = "^1.1.1"
web3 = "^5.31.3"
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
